2018/requirements.txt
blist==1.3.6
cycler==0.10.0
decorator==4.3.0
kiwisolver==1.0.1
matplotlib==3.0.2
networkx==2.2
numpy==1.15.4
pycairo==1.18.0
pyparsing==2.3.0
python-dateutil==2.7.5
python-igraph==0.7.1.post6
six==1.12.0
